% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{analyze_all}
\alias{analyze_all}
\title{Analyze all .RDS files (output from type_all_packages) in a specified dir.
NOTE: only caring about typeof at this time}
\usage{
analyze_all(path)
}
\description{
Analyze all .RDS files (output from type_all_packages) in a specified dir.
NOTE: only caring about typeof at this time
}
