% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{analyze_attributes}
\alias{analyze_attributes}
\title{analyze_attributes :: look @ & aggregate attribute information over list of files}
\usage{
analyze_attributes(lof, path_to_dir = "", pname = "")
}
\arguments{
\item{lof}{the files to read (list.files() of the /tmp/... stuff)}

\item{path_to_dir}{the path to the directory}

\item{pname}{the name of the package}
}
\description{
analyze_attributes :: look @ & aggregate attribute information over list of files
}
