% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{analyze_type_information}
\alias{analyze_type_information}
\title{analyze_type_information takes a tally and lists some things about it:
[1] how many easy monomorphic, hard monomorphic, and polymorphic functions
    are there
[2] for easy monomorphic functions, suggestions for annotation}
\usage{
analyze_type_information(tally)
}
\arguments{
\item{tally}{the tally to analyze, use readRDS on type_from_package output}
}
\description{
analyze_type_information takes a tally and lists some things about it:
[1] how many easy monomorphic, hard monomorphic, and polymorphic functions
    are there
[2] for easy monomorphic functions, suggestions for annotation
}
