% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testit.R
\name{find_singletons_in_old_trace}
\alias{find_singletons_in_old_trace}
\title{Temporary function for compatibility with old, already saved tallies.
Intended to bring old data "up to speed" to stay on top of changes to
the generation process.}
\usage{
find_singletons_in_old_trace(tally)
}
\description{
Temporary function for compatibility with old, already saved tallies.
Intended to bring old data "up to speed" to stay on top of changes to
the generation process.
}
