% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{get_all_for_kind}
\alias{get_all_for_kind}
\title{Get all trace results for a particular kind of polymorphism.}
\usage{
get_all_for_kind(lot, kind)
}
\arguments{
\item{lot}{result from analyze_all}
}
\description{
Get all trace results for a particular kind of polymorphism.
}
