% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{give_only_polymorphic}
\alias{give_only_polymorphic}
\title{Give Only Polymorphic :: take a data.frame (from transform_results) and give
  back a data frame where the `singleton` and `monomorphic` fields are gone.}
\usage{
give_only_polymorphic(df)
}
\arguments{
\item{df}{result from transform_results}
}
\description{
Give Only Polymorphic :: take a data.frame (from transform_results) and give
  back a data frame where the `singleton` and `monomorphic` fields are gone.
}
