% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel-functions.R
\name{install_in_parallel}
\alias{install_in_parallel}
\title{Function to install multiple CRAN packages using parallel instances.
Parameters:}
\usage{
install_in_parallel(packages, parallel.installs)
}
\arguments{
\item{packages}{- Packages to be installed in vector}

\item{parallel.installs}{- Number of parallel installs}
}
\description{
Function to install multiple CRAN packages using parallel instances.
Parameters:
}
