% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{plot_bar}
\alias{plot_bar}
\title{Plot a bar chart: uses results from transform_results and give_only_polymorphic}
\usage{
plot_bar(df)
}
\description{
Plot a bar chart: uses results from transform_results and give_only_polymorphic
}
