% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{plot_pie}
\alias{plot_pie}
\title{Plot a pie chart: uses results from transform_results and give_only_polymorphic}
\usage{
plot_pie(df)
}
\description{
Plot a pie chart: uses results from transform_results and give_only_polymorphic
}
