% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{prep_for_parallel}
\alias{prep_for_parallel}
\title{Function which takes a list of package_names (for big parallel call to
type_a_package), and prepares for the call.}
\usage{
prep_for_parallel(num_cores = 1)
}
\description{
Function which takes a list of package_names (for big parallel call to
type_a_package), and prepares for the call.
}
