% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{print_tally_results}
\alias{print_tally_results}
\title{print_tally_results will display the type, class, mode, and storage.mode
of all arguments of all functions in a specified tally. It is intended to
be human-readable.}
\usage{
print_tally_results(tally)
}
\arguments{
\item{tally}{the tally to print}
}
\description{
print_tally_results will display the type, class, mode, and storage.mode
of all arguments of all functions in a specified tally. It is intended to
be human-readable.
}
