% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testit.R
\name{reprocess_all_tmps}
\alias{reprocess_all_tmps}
\title{Small function to reprocess existing tmp files. Use this if you made
any modifications to the result generating functions, and would like
to have your result_XYZ files reflect them.}
\usage{
reprocess_all_tmps()
}
\description{
Small function to reprocess existing tmp files. Use this if you made
any modifications to the result generating functions, and would like
to have your result_XYZ files reflect them.
}
