% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{transform_results}
\alias{transform_results}
\title{Transform :: take a count result from an analysis, and transform it
  to a data.frame that can be used to plot.}
\usage{
transform_results(ana_res)
}
\arguments{
\item{ana_res}{the *whole* result of a call to analyze_all}
}
\description{
Transform :: take a count result from an analysis, and transform it
  to a data.frame that can be used to plot.
}
