% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{type_a_package}
\alias{type_a_package}
\title{Type just one package. Use this with foreach ... %dopar% if you want
parallelism.}
\usage{
type_a_package(package_name, clean = FALSE, skipgen = FALSE,
  quiet = TRUE)
}
\description{
Type just one package. Use this with foreach ... %dopar% if you want
parallelism.
}
