% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{type_all_packages}
\alias{type_all_packages}
\title{Extract type information from many packages, and consolidate information.
e.g. how many functions have an argument with more than one type?}
\usage{
type_all_packages(package_names, clean = FALSE, skipgen = FALSE,
  use_rev_deps = FALSE)
}
\arguments{
\item{package_names}{the list of package names to type.}

\item{clean}{should we clean up the type_res directory?}
}
\description{
Extract type information from many packages, and consolidate information.
e.g. how many functions have an argument with more than one type?
}
