% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{type_from_package}
\alias{type_from_package}
\title{Extract function argument type information from a specified package.}
\usage{
type_from_package(pkgs_to_trace, pkgs_to_run = pkgs_to_trace,
  types = c("examples", "tests", "vignettes", "all"), filter = NULL,
  quiet = TRUE, skipgen = FALSE, ...)
}
\description{
Extract function argument type information from a specified package.
}
