% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testit.R
\name{type_trace_all_tally}
\alias{type_trace_all_tally}
\title{tally all traces pointed to by the specified file names}
\usage{
type_trace_all_tally(file_names, path_to_dir)
}
\description{
data format :
res[[x]][y] : res is the result
						  x is the test #
							y is param name
}
