% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testit.R
\name{type_trace_args_tally}
\alias{type_trace_args_tally}
\title{Tally up a trace.}
\usage{
type_trace_args_tally(a_trace)
}
\arguments{
\item{a_trace}{the trace to analyze}

\item{tally}{the tally to add the analysis to}
}
\description{
Tally up a trace.
}
