% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testit.R
\name{update_all_traces}
\alias{update_all_traces}
\title{Temporary function to update old traces for compliance.}
\usage{
update_all_traces(path)
}
\description{
Temporary function to update old traces for compliance.
}
