% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{analyze_all}
\alias{analyze_all}
\title{Analyze all .RDS files (output from type_all_packages) in a specified dir.}
\usage{
analyze_all(path, type = "type", display = "some", per_arg = FALSE)
}
\arguments{
\item{path}{path to the results of type_all_packages (this should be something
of the form */type_res)}

\item{type}{the type of info you want. currently either "type" or "class"}
}
\description{
Analyze all .RDS files (output from type_all_packages) in a specified dir.
}
