% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typethat.R
\name{analyze_type_information}
\alias{analyze_type_information}
\title{analyze_type_information takes a tally and lists some things about it:
[1] how many easy monomorphic, hard monomorphic, and polymorphic functions
    are there
[2] for easy monomorphic functions, suggestions for annotation}
\usage{
analyze_type_information(tally, type = "type")
}
\arguments{
\item{tally}{the tally to analyze, use readRDS on type_from_package output}

\item{type}{the type of info to get. either "type" or "class".}
}
\description{
Types of ``morphicity'':
monomorphic
optional monomorphic
simple polymorphic
optional simple polymorphic
complex polymorphic
optional polymorphic
complex polymorphic with other optional argument
}
